Fixed a bug in user_food grams_per_nutrient
Previously it raised 'unsupported' when a serving size
was neither mass nor had density. Now it returns serving
size as a non-mass unit, which makes it a poorly named method
but is more accurate to real life, specifically when something
only has volume as a realistic measurement (ex. soy sauce).