Fix capybara test to identify itself as iphone
Capybara doesn't have a convenient way to specify headers to send like
webrat does so we have to hack into Rack test to get it to correctly
identify itself as an iPhone. Unfortunately this doesn't seem to work
on (some?) redirects from the application so the cucumber test has to
take an extra step in one case to make sure it is rendering the mobile
version of the site.