Fix possessive bug on Recipes list page
When viewing the recipe list for a specific user the heading of the
page displays the user's name in the possessive case. It was doing
this by appending an 's' to the name. Fix it so it properly appending
and apostraphe and an 's' to names that don't end with 's', and just
an apostraphe to ones that do.